# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Cookiebot backend API","description":"Cookiebot backend services","license":{"name":"Apache 2.0"},"version":"1.2.1"},"externalDocs":{"description":"Cookiebot backend README","url":"https://github.com/MekhyW/COOKIEBOT-backend"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"User name and password":[]}],"tags":[{"name":"BFF groups","description":"BFF telegram groups API"},{"name":"groups","description":"telegram groups API"},{"name":"BFF bot config","description":"BFF telegram bot configuration API"}],"paths":{"/welcomes/{id}":{"get":{"tags":["welcome-resource"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Welcome"}}}}}},"put":{"tags":["welcome-resource"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Welcome"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["welcome-resource"],"operationId":"insert","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Welcome"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Welcome"}}}}}},"delete":{"tags":["welcome-resource"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/users/{id}":{"get":{"tags":["user-resource"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-resource"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-resource"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/rules/{id}":{"get":{"tags":["rule-resource"],"operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rule"}}}}}},"put":{"tags":["rule-resource"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rule"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["rule-resource"],"operationId":"insert_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rule"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rule"}}}}}},"delete":{"tags":["rule-resource"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/registers/{id}/users":{"get":{"tags":["register-resource"],"operationId":"findUsers","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRegister"}}}}}}},"put":{"tags":["register-resource"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["register-resource"],"operationId":"insertUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["register-resource"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/raffles/{name}":{"get":{"tags":["raffle-resource"],"operationId":"findByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Raffle"}}}}}},"put":{"tags":["raffle-resource"],"operationId":"update_3","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Raffle"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["raffle-resource"],"operationId":"insert_6","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Raffle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Raffle"}}}}}},"delete":{"tags":["raffle-resource"],"operationId":"delete_4","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/raffles/{name}/participants":{"get":{"tags":["raffle-resource"],"operationId":"findParticipants","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RaffleParticipant"}}}}}}},"put":{"tags":["raffle-resource"],"operationId":"updateParticipant","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaffleParticipant"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["raffle-resource"],"operationId":"insertParticipant","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaffleParticipant"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["raffle-resource"],"operationId":"deleteParticipant","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaffleParticipant"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/groups/{id}":{"get":{"tags":["groups"],"summary":"Find a group by id","description":"Find a telegram group by id","operationId":"findByGroupId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Group does not exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Group"}}}},"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}}}},"put":{"tags":["groups"],"summary":"Upsert a group","description":"Updates a telegram group or create it if it does not exists","operationId":"upsert","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}}}},"post":{"tags":["groups"],"summary":"Create a new group","description":"Create a new telegram group","operationId":"insert_7","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}},"required":true},"responses":{"422":{"description":"Group with the same ID already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Group"}}}},"201":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}}}},"delete":{"tags":["groups"],"summary":"Delete a group","description":"Deletes a telegram group","operationId":"delete_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"404":{"description":"Group does not exists"}}}},"/groups/{id}/admins":{"get":{"tags":["groups"],"summary":"List admins of a group","description":"List the admin users of a given group ID","operationId":"findAdmins","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"404":{"description":"Group does not exists","content":{"*/*":{"schema":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}}}}},"put":{"tags":["groups"],"summary":"Set group admin","description":"Set the given admin users to the telegram group removing users that are not in the list.","operationId":"updateAdmins","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"404":{"description":"Group does not exists"}}},"post":{"tags":["groups"],"summary":"Add admins to group","description":"Adds the given admin users to the telegram group keeping the existing ones.","operationId":"insertAdmins","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"404":{"description":"Group does not exists"}}},"delete":{"tags":["groups"],"summary":"Remove group admins","description":"Remove the given admin users to the telegram group keeping the existing ones.","operationId":"deleteAdmins","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"userIds","in":"query","required":true,"schema":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"404":{"description":"Group does not exists"}}}},"/events/{id}":{"get":{"tags":["event-resource"],"operationId":"findById_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Event"}}}}}},"put":{"tags":["event-resource"],"operationId":"update_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Event"}}}}}},"delete":{"tags":["event-resource"],"operationId":"delete_6","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/configs/{id}":{"get":{"tags":["config-resource"],"operationId":"findById_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Config"}}}}}},"put":{"tags":["config-resource"],"operationId":"update_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["config-resource"],"operationId":"insert_9","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Config"}}}}}},"delete":{"tags":["config-resource"],"operationId":"delete_7","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/bff/group/{groupId}/config":{"get":{"tags":["BFF bot config"],"summary":"Get configuration","description":"Get group bot configuration","operationId":"getGroupConfig","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Group does not exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Config"}}}},"403":{"description":"User does not have permission to access bot configuration","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Config"}}}},"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}}}}},"put":{"tags":["BFF bot config"],"summary":"Update configuration","description":"Update bot configuration for group","operationId":"updateGroupConfig","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}},"required":true},"responses":{"404":{"description":"Group does not exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Config"}}}},"403":{"description":"User does not have permission to update bot configuration","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Config"}}}},"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}}}}}},"/users":{"get":{"tags":["user-resource"],"operationId":"findAll_1","parameters":[{"name":"username","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-resource"],"operationId":"insert_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/stickerdatabase":{"get":{"tags":["sticker-database-resource"],"operationId":"findAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StickerDatabase"}}}}}},"post":{"tags":["sticker-database-resource"],"operationId":"insert_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StickerDatabase"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StickerDatabase"}}}}}}},"/registers/{id}":{"get":{"tags":["register-resource"],"operationId":"findById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Register"}}}}}},"post":{"tags":["register-resource"],"operationId":"insert_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Register"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Register"}}}}}},"delete":{"tags":["register-resource"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/randomdatabase":{"get":{"tags":["random-database-resource"],"operationId":"findAll_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RandomDatabase"}}}}}},"post":{"tags":["random-database-resource"],"operationId":"insert_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RandomDatabase"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RandomDatabase"}}}}}}},"/events":{"get":{"tags":["event-resource"],"operationId":"findAll_8","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Event"}}}}}}},"post":{"tags":["event-resource"],"operationId":"insert_8","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Event"}}}}}}},"/blacklist/{id}":{"get":{"tags":["blacklist-resource"],"operationId":"findById_6","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blacklist"}}}}}},"post":{"tags":["blacklist-resource"],"operationId":"insert_10","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rule"}}}}}},"delete":{"tags":["blacklist-resource"],"operationId":"delete_8","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/bff/events":{"get":{"tags":["events-controller"],"operationId":"getEvents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","properties":{"parallel":{"type":"boolean"}}}}}}}},"post":{"tags":["events-controller"],"operationId":"createEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Event"}}}}}}},"/welcomes":{"get":{"tags":["welcome-resource"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Welcome"}}}}}}}},"/rules":{"get":{"tags":["rule-resource"],"operationId":"findAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Rule"}}}}}}}},"/registers":{"get":{"tags":["register-resource"],"operationId":"findAll_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Register"}}}}}}}},"/raffles":{"get":{"tags":["raffle-resource"],"operationId":"findAll_6","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Raffle"}}}}}}}},"/groups":{"get":{"tags":["groups"],"summary":"Get all groups","description":"Get all registered telegram groups","operationId":"findAll_7","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Group"}}}}}}}},"/configs":{"get":{"tags":["config-resource"],"operationId":"findAll_9","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Config"}}}}}}}},"/blacklist":{"get":{"tags":["blacklist-resource"],"operationId":"findAll_10","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Blacklist"}}}}}}}},"/bff/my-groups":{"get":{"tags":["BFF groups"],"summary":"Get user groups","description":"Get groups where user can edit","operationId":"getMyGroups","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Group"}}}}}}}}},"components":{"schemas":{"Welcome":{"type":"object","properties":{"id":{"type":"string"},"message":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"languageCode":{"type":"string"},"birthdate":{"type":"string","format":"date"}}},"Rule":{"type":"object","properties":{"id":{"type":"string"},"rules":{"type":"string"}}},"UserRegister":{"type":"object","properties":{"user":{"type":"string"},"date":{"type":"string"},"accountId":{"type":"string"}}},"Raffle":{"type":"object","properties":{"name":{"type":"string"},"award":{"type":"string"},"deadline":{"type":"string"},"participants":{"type":"array","items":{"$ref":"#/components/schemas/RaffleParticipant"}}}},"RaffleParticipant":{"type":"object","properties":{"user":{"type":"string"}}},"Group":{"type":"object","properties":{"groupId":{"type":"string"},"name":{"type":"string"},"imageUrl":{"type":"string"},"adminUsers":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"Event":{"required":["city","country","description","endDate","groupId","locationName","name","price","startDate","state"],"type":"object","properties":{"id":{"type":"string"},"groupId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"locationName":{"type":"string"},"coverImageUrl":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"},"price":{"type":"integer","format":"int32"},"active":{"type":"boolean"}}},"Config":{"type":"object","properties":{"id":{"type":"string"},"furbots":{"type":"boolean"},"stickerSpamLimit":{"type":"string"},"timeWithoutSendingImages":{"type":"integer","format":"int32"},"timeCaptcha":{"type":"integer","format":"int32"},"functionsFun":{"type":"boolean"},"functionsUtility":{"type":"boolean"},"sfw":{"type":"boolean"},"language":{"type":"string"},"publisherPost":{"type":"boolean"},"publisherAsk":{"type":"boolean"},"publisherMembersOnly":{"type":"boolean"},"threadPosts":{"type":"string"},"maxPosts":{"type":"integer","format":"int32"}}},"StickerDatabase":{"type":"object","properties":{"id":{"type":"string"}}},"Register":{"type":"object","properties":{"id":{"type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/UserRegister"}}}},"RandomDatabase":{"type":"object","properties":{"id":{"type":"string"},"idMessage":{"type":"string"},"idMedia":{"type":"string"}}},"Blacklist":{"type":"object","properties":{"id":{"type":"string"}}}},"securitySchemes":{"User name and password":{"type":"http","scheme":"basic"},"JWT token":{"type":"http","scheme":"bearer"}}}}